refactor: lineService의 비지니스 로직 domain으로 옮기기
