+++ build.gradle
@@ -25,6 +25,10 @@ dependencies {
 	// log
 	implementation 'net.rakugakibox.spring.boot:logback-access-spring-boot-starter:2.7.1'
 
+	// lombok
+	compileOnly 'org.projectlombok:lombok'
+	annotationProcessor 'org.projectlombok:lombok'
+
 	// jgraph
 	implementation 'org.jgrapht:jgrapht-core:1.0.1'
